MCTIERS-Bot-main/requirements.txt
discord.py
python-dotenv
mysql-connector-python
aiohttp
discord-html-transcripts
